convert to flat storage
