--- src/services/fileStorage/utils/leadingSlashHelper.js
@@ -1,4 +1,0 @@
-module.exports  = path => {
-	if (path[0] === '/') path = path.substring(1);
-	return path;
-};
